requirements.txt
# Reddit API
praw>=7.7.0

# Sentiment Analysis
vaderSentiment>=3.3.2

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Dashboard
streamlit>=1.28.0
plotly>=5.17.0

# Text visuals
wordcloud>=1.9.3

# Utilities
python-dotenv>=1.0.0

# Scheduling (for background data collection)
APScheduler>=3.10.0

# Word clouds (optional)
wordcloud>=1.9.0
